+++ src/actions/ActionsMap.php
@@ -59,13 +59,62 @@ class ActionsMap extends BaseObject
     }
 
     /**
-     * Run action for a node if any
+     * Run action for a node and apply its result to node
      *
-     * Runs action for a node if action is defined.
+     * Runs action for a node if action is defined. Returns whether action was found
+     * and its result was made. Result of the action will be applied with `make()`
+     * only when it is not `null`. So you can call `make()` manually inside action handler
+     * under some conditions.
+     *
+     * To apply `null` to node with actions the only way is to call `make(null)` inside
+     * action handler. It should not be a problem since default `made()` value on nodes
+     * is `null`.
      * @param TreeNodeInterface $node Subject node
-     * @return mixed Value returned from action or `null`.
+     * @return bool|null Returns `null` when no action defined. Returns `false` when
+     * action result `null` and so it was not applied. Returns `true` wher action result
+     * was not `null` and it was applied.
+     * @since 1.3.2
+     */
+    public function applyToNode($node)
+    {
+        $action = $this->getAction($node);
+        if (null === $action) {
+            return null;
+        }
+        $value = $this->runActionHandler($action, $node);
+        if (null === $value) {
+            return false;
+        }
+        $node->make($value);
+        return true;
+    }
+
+    /**
+     * Run action for a node and returns its result
+     *
+     * Runs action for a node if action is defined. Returns result of the action.
+     * @param TreeNodeInterface $node Subject node
+     * @return mixed Value returned from action or `null` if no action.
      */
     public function runForNode($node)
+    {
+        // REFACT: unused internally - deprecate or improve public interface
+
+        $action = $this->getAction($node);
+        if (null === $action) {
+            return null;
+        }
+
+        return $this->runActionHandler($action, $node);
+    }
+
+    /**
+     * Get action handler for node
+     * @param TreeNodeInterface $node Subject node
+     * @return callable|string|null Action handler or `null`.
+     * @since 1.3.2
+     */
+    private function getAction($node)
     {
         $name = $node->getNodeName();
         $tag = $node->getNodeTag();
@@ -73,12 +122,23 @@ class ActionsMap extends BaseObject
             $name .= '(' . $tag . ')';
         }
 
+        // REFACT: minimal PHP >= 7.0: $var ?? null
         if (!isset($this->actions[$name])) {
             return null;
         }
 
-        $action = $this->actions[$name];
+        return $this->actions[$name];
+    }
 
+    /**
+     * Run action and return its result
+     * @param callable|string $action Action from the map
+     * @param TreeNodeInterface $node Subject node
+     * @return mixed Result of the action call
+     * @since 1.3.2
+     */
+    private function runActionHandler($action, $node)
+    {
         if (is_string($action) && isset(self::$COMMANDS[$action])) {
             /** @var CommandInterface $class Just a class name, not an instance */
             $class = self::$COMMANDS[$action];