+++ tests/unit/actions/ActionsMapTest.php
@@ -9,6 +9,32 @@ use VovanVE\parser\tree\NonTerminal;
 
 class ActionsMapTest extends BaseTestCase
 {
+    public function testApplyToNode()
+    {
+        $map = new ActionsMap([
+            'foo' => function (Token $foo) {
+                return $foo->getContent();
+            },
+            'Bar' => function (TreeNodeInterface $bar, TreeNodeInterface $foo) {
+                // apply instead of return
+                $bar->make('[' . $foo->made() . ']');
+                // so return null
+            },
+        ]);
+
+        $foo = new Token('foo', 'lorem ipsum');
+        $this->assertTrue($map->applyToNode($foo));
+
+        $bar = new NonTerminal('Bar', [$foo]);
+        $this->assertFalse($map->applyToNode($bar));
+        $this->assertEquals('[lorem ipsum]', $bar->made());
+
+        $baz = new Token('baz', 'dolor');
+        $baz->make(42);
+        $this->assertNull($map->applyToNode($baz));
+        $this->assertSame(42, $baz->made());
+    }
+
     public function testRunForNode()
     {
         $map = new ActionsMap([